+++ forge/CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Bug Fixes
 
 - (Not really a bug) Corrected naming of parameters in ItemPropertyWrapperTemplates.
+- Updated PropertyWrapperContainer#MAPPED_PROPERTY_WRAPPERS to properly compare wrapper/supplier values in its equality check.
+- Updated PropertyWrapper#getMappedPropertyWrappers() HashStrategy implementation to match PropertyWrapperContainer#MAPPED_PROPERTY_WRAPPERS.
 
 ## New Features